third_party/blink/web_tests/platform/generic/fast/events/touch/touch-coords-in-zoom-and-scroll-expected.txt

Base
PASS event.touches[0].clientX is 100
PASS event.touches[0].clientY is 100
PASS event.touches[0].pageX is 100
PASS event.touches[0].pageY is 100

Just zoomed
PASS event.touches[0].clientX is within 0.00001 of 83.33333
PASS event.touches[0].clientY is within 0.00001 of 83.33333
PASS event.touches[0].pageX is within 0.00001 of 83.33333
PASS event.touches[0].pageY is within 0.00001 of 83.33333

Just scrolled
PASS event.touches[0].clientX is 100
PASS event.touches[0].clientY is 100
PASS event.touches[0].pageX is 150
PASS event.touches[0].pageY is 150

Zoomed and scrolled
PASS event.touches[0].clientX is within 0.00001 of 83.33333
PASS event.touches[0].clientY is within 0.00001 of 83.33333
PASS event.touches[0].pageX is within 0.00001 of 133.33333
PASS event.touches[0].pageY is within 0.00001 of 133.33333
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE
